requirements.txt
Django==3.2.6
celery==5.0.2
psycopg2-binary==2.8.5
djangorestframework>=3.13.1
django-cors-headers>=3.12.0
django-rest-swagger>=2.2.0
black
pylint